iiod: Move read_all/write_all to new file
These functions are used by both IIOD and tinyiiod, and there's no
convenient place to place them; so move them to a new rw.c file.
