+++ iiod/interpreter.c
@@ -195,40 +195,6 @@ static ssize_t writefd_io(struct parser_pdata *pdata, const void *src, size_t le
 	return ret;
 }
 
-ssize_t write_all(struct parser_pdata *pdata, const void *src, size_t len)
-{
-	uintptr_t ptr = (uintptr_t) src;
-
-	while (len) {
-		ssize_t ret = pdata->writefd(pdata, (void *) ptr, len);
-		if (ret < 0)
-			return ret;
-		if (!ret)
-			return -EPIPE;
-		ptr += ret;
-		len -= ret;
-	}
-
-	return ptr - (uintptr_t) src;
-}
-
-ssize_t read_all(struct parser_pdata *pdata, void *dst, size_t len)
-{
-	uintptr_t ptr = (uintptr_t) dst;
-
-	while (len) {
-		ssize_t ret = pdata->readfd(pdata, (void *) ptr, len);
-		if (ret < 0)
-			return ret;
-		if (!ret)
-			return -EPIPE;
-		ptr += ret;
-		len -= ret;
-	}
-
-	return ptr - (uintptr_t) dst;
-}
-
 void interpreter(struct iio_context *ctx, int fd_in, int fd_out,
 		 bool is_socket, bool is_usb,
 		 struct thread_pool *pool, const void *xml_zstd,
